cormoran/fields.py
# -*- coding: utf-8 -*-
#
# Cormoran is a fast and lightweight persistence framework.
# Copyright (C) 2012 Jaime Gil de Sagredo Luna <jaimegildesagredo@gmail.com>
#
# This program is free software: you can redistribute it and/or modify
# it under the terms of the GNU General Public License as published by
# the Free Software Foundation, either version 3 of the License, or
# (at your option) any later version.
#
# This program is distributed in the hope that it will be useful,
# but WITHOUT ANY WARRANTY; without even the implied warranty of
# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
# GNU General Public License for more details.
#
# You should have received a copy of the GNU General Public License
# along with this program.  If not, see <http://www.gnu.org/licenses/>.


class FieldError(Exception):
    pass


class BaseField(object):
    def __init__(self, name=None, default=None):
        if self.__class__ is BaseField:
            raise FieldError()
        self.name = name
        self.default = default

    def __get__(self, instance, owner):
        if instance is None:
            return self

        return self.default


class StringField(BaseField):
    pass

